Rework de la vistas simples

+++ db/migrate/20190919142202_create_sizes.rb
@@ -1,7 +1,7 @@
 class CreateSizes < ActiveRecord::Migration[5.2]
   def change
     create_table :sizes do |t|
-      t.decimal :number
+      t.integer :number
       t.decimal :us
       t.decimal :cms
